Extracted sm64 map logic from mario.h
It now resides on levelsm64.h. This is a WIP.
