+++ src/format.h
@@ -7029,4 +7029,8 @@ namespace TR {
     }; // struct Level
 }
 
+struct IController {
+    virtual void getFloorInfo(int roomIndex, const vec3 &pos, TR::Level::FloorInfo &info) const {}
+};
+
 #endif
